BUILD: Fix build method for plugin
I'm not sur of this way. Maybe main binary and demo plugin should be
stored in différent repository than library.

This is for compatibility between binary and module. While the main
program and the plugin uses common libraries, if the main program
is compiled as part of these libraries, go detect a version problem.